apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Urban tree grant application backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "python-multipart",
    "shared-models",
    "database",
    "authentication",
    "storage",
    "email",
    "Pillow",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "httpx",
    "mypy",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "def __repr__", "raise NotImplementedError"]

[tool.uv.sources]
shared-models = { workspace = true }
database = { workspace = true }
authentication = { workspace = true }
storage = { workspace = true }
email = { workspace = true }
